Fix inversed colour temperature.
The colour temperature was inversed so when you picked warm in the ui the lightbulb would turn cold and vice versa. This is because the range of 140-500 in homekit is inverse to the 0-1000 range of the lightbulb, in fact the values 140-500 are micro-reciprocal degrees (mired), which is 1,000,000 divided by the color temperature in kelvins.

I've adjusted the range conversions for colour temperature to convert to and from the range 2000K-14285K which when converted to mireds corresponds to about 70-500 inversed.

I've also added a little trick extending the colour temperature characteristics range to 70-500 in order to show a ui that has a nice orange-blue gradient. Hence I used the values 70-500 in the calculations as well.